+++ CHANGELOG.md
@@ -21,6 +21,7 @@ All notable changes to this project will be documented in this file.
 - Update placeholder handling
 
 ### Fixed
+- phone-codes.js is missing when installing with bower #937
 - Repeat function doesn't work for dynamic masks #960
 - Provide convenient method to unmask value #929
 - Min value doesn't work with allowMinus #951